+++ package.json
@@ -6,7 +6,7 @@
     "@babel/core": "^7.16.0",
     "@emotion/react": "^11.9.0",
     "@emotion/styled": "^11.8.1",
-    "@fontsource/roboto": "^4.5.5",
+    "@fontsource/roboto": "4.5.5",
     "@material-ui/core": "^4.11.4",
     "@material-ui/icons": "^4.11.2",
     "@material-ui/lab": "^4.0.0-alpha.60",
