+++ src/deploy/index.js
@@ -1,0 +1,11 @@
+import React from 'react';
+import FilterItem from './filter';
+
+export default function Explore(props) {
+    console.log('Explore')
+    return (
+        <div>
+           <FilterItem  />
+        </div>
+    )
+}
