+++ src/fameScan/index.js
@@ -1,0 +1,12 @@
+import React from 'react';
+
+export default function Market(props) {
+    console.log('FAMEScan')
+    return (
+        <div>
+            <p>
+                FAMEScan
+            </p>
+        </div>
+    )
+}
