+++ src/router/index.js
@@ -2,11 +2,15 @@ import { Route, Routes } from "react-router-dom";
 import React, { Suspense } from "react";
 
 const Home = React.lazy(() => import("@/home"));
-const Explore = React.lazy(() => import("@/explore"));
+const Deploy = React.lazy(() => import("@/deploy"));
+const FameScan = React.lazy(() => import("@/fameScan"));
+const Market = React.lazy(() => import("@/market"));
 const RouterPage = ({ initialRoute }) => (
     <Routes>
       <Route path="/" element={<Home />} />
-      <Route path="/explore" element={<Explore />} />
+      <Route path="/deploy" element={<Deploy />} />
+      <Route path="/fameScan" element={<FameScan />} />
+      <Route path="/market" element={<Market />} />
     </Routes>
 );
 export default RouterPage;
